# pkg/infrastructure/rest/controllers/diner/Responses.go
// Package diner contains the diner controller
package diner

// MessageResponse is a struct that contains the response body for the message
type MessageResponse struct {
	Message string `json:"message"`
}